Separated the utils and constants files to reflect functions and constants
